attack/requirements.txt
flask==2.0.1
flask-sock==0.7.0
kafka-python==2.0.2
gunicorn==20.1.0
gevent==21.8.0
gevent-websocket==0.10.1
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
tqdm>=4.65.0